requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
opencv-python>=4.8.1.78
numpy>=1.26.0
ultralytics>=8.0.196
PyPDF2>=3.0.1
pdf2image>=1.16.3
pydantic>=2.5.0
pydantic-settings>=2.1.0
shapely>=2.0.2
pandas>=2.1.3
aiofiles>=23.2.1
python-dotenv>=1.0.0
pytest>=7.4.3

